+++ src/utilis/utils.js
@@ -1,0 +1,26 @@
+export const checkIputs = (randomData, inputValues) => {
+  let correctLocation = 0;
+  let correctValue = 0;
+
+  let randomFruitCopy = randomData.slice();
+  let droppedItemsCopy = inputValues.slice();
+
+  randomFruitCopy.forEach((fruit, i) => {
+    if (droppedItemsCopy[i] === fruit) {
+      correctLocation++;
+      correctValue++;
+      droppedItemsCopy[i] = randomFruitCopy[i] = null;
+    }
+  });
+
+  droppedItemsCopy.forEach((input, i) => {
+    if (input === null) return;
+    let foundIdx = randomFruitCopy.indexOf(input);
+    if (foundIdx > -1) {
+      correctValue++;
+      randomFruitCopy[foundIdx] = null;
+    }
+  });
+
+  return { correctValue, correctLocation };
+};
